move code to mgmt command
